Trait harness::Module

pub trait Module {
    fn run<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        scheduler: &'life1 JobScheduler
    ) -> Pin<Box<dyn Future<Output = Result<Option<Heart>, BoxedError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn pre_startup<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }

    fn post_shutdown<'life0, 'async_trait>(
        &'life0 mut self, 
        termination_reason: ModuleTerminationReason
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
}

Executable module

Required methods

fn run<'life0, 'life1, 'async_trait>(
&'life0 mut self,
scheduler: &'life1 JobScheduler
) -> Pin<Box<dyn Future<Output = Result<Option<Heart>, BoxedError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Core run loop of the module

When the function returns Some(_) the death of the returned Heart is awaited before calling the shutdown hook. Useful for situations where you dispatch background jobs in the run loop but want to hand-off the program lifecycle management.

Returning None results in the program entering a shutdown state and calling the pre_shutdown hook.

Provided methods

fn pre_startup<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

Executed before running the core loop

fn post_shutdown<'life0, 'async_trait>(
&'life0 mut self,
termination_reason: ModuleTerminationReason
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Send + 'async_trait,

Shutdown hook executed after the core loop and all associated jobs have terminated